refactor(download): refactor expression evaluations
